[Q2] Rename file to match validate.sh's spec
