+++ hw2/emjames/q2/_vector.h
