+++ hw2/emjames/q2/test_calcAngle.py
@@ -1,11 +1,11 @@
-import calcAngle
+import _vector 
 import pytest
 
 def test_angle1():
-    angle = calcAngle.findAngle2d("3,4", "4,3")
+    angle = _vector.findAngle2d("3,4", "4,3")
     assert angle == 0.96
 
 def test_angle2():
-    angle = calcAngle.findAngle2d("7,1", "5,5")
+    angle = _vector.findAngle2d("7,1", "5,5")
     assert round(angle, 2) == 0.80
